Add files via upload
This SQL Project was a group project completed for Data Warehousing class. In the project we pulled data files from different sources and and utilize them to create a Data Lake structure that we can then query to produce visualizations. We utilized databricks for both the Data Lake development and the queries/visualizations.